Major update for Scoping. rescoped several classes to Discipline.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@3821 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/TaxonAttachment.java
@@ -201,6 +201,6 @@ public class TaxonAttachment extends DataModelObjBase implements ObjectAttachmen
      */
     public int compareTo(TaxonAttachment o)
     {
-        return ((Integer)getOrderIndex()).compareTo((Integer)o.getOrderIndex());
+        return ((Integer)getOrderIndex()).compareTo(o.getOrderIndex());
     }
 }
